+++ kpresenter/kppolygonobject.h
@@ -66,8 +66,8 @@ public:
     void setPolygonSettings( bool _checkConcavePolygon, int _cornersValue, int _sharpnessValue );
     void getPolygonSettings( bool *_checkConcavePolygon, int *_cornersValue, int *_sharpnessValue );
 
-    virtual QDomDocumentFragment save( QDomDocument& doc, int offset );
-    virtual int load( const QDomElement &element );
+    virtual QDomDocumentFragment save( QDomDocument& doc, double offset );
+    virtual double load( const QDomElement &element );
 
 protected:
     virtual void paint( QPainter *_painter,KoZoomHandler*_zoomHandler, bool drawingShadow );
