+++ kpresenter/kptextobject.h
@@ -78,8 +78,8 @@ public:
 
     void setShadowParameter(int _distance,ShadowDirection _direction,QColor _color);
 
-    virtual QDomDocumentFragment save( QDomDocument& doc, int offset );
-    virtual int load(const QDomElement &element);
+    virtual QDomDocumentFragment save( QDomDocument& doc, double offset );
+    virtual double load(const QDomElement &element);
 
     virtual void draw( QPainter *_painter, KoZoomHandler*_zoomHandler, bool drawSelection );
     void draw( QPainter *_painter, KoZoomHandler*_zoomHandler, bool onlyChanged, QTextCursor* cursor, bool resetChanged, bool drawSelection );
